[todo] Add todo package
This package declares a List type which represent a list of todos.

There's also the unexported item type which represents a todo.

Since this type is not exported, methods are defined on List to interact with todos.

Tests are also written for public Add, Complete, Delete, Get and Save methods.