Replace ReadAll with Read to lower the CPU and memory usage
After profiling, it can be seen that the GC is simply doing too much work.
This happens because we're allocating the whole data into one slice.

To fix this, csv files are read one record at a time and csvReader.ReuseRecord is used to not allocate new slices for each record.

These conclusions are made after profiling the CPU and mem usage of the CLI by adding a simple benchmark test which tests the CLI based on the csv under /benchmark directory.
The pprof files are diagnosed by using go tool.