Add a new step for gofmt which utilizes stdout
A new step to the goci pipeline is added for formatting the given project using gofmt.

However, this step is different than previous ones in a way that gofmt does not return an error if it fails, but writes the errors on stdout.

Therefore, a new step called exceptionStep.go is added to the project, and a new executer interface is used to maintain both step and exceptionStep within the goci pipeline.