Add subcommands for host such as add, delete and list
The host business logic which was written in scan package is connected with Cobra via the methods below:

- addAction
- removeAction
- deleteAction

Cobra parses the flags and passes them to the functions above to do operations on hosts.

Unit tests are added for add, remove and delete actions along with an integration test which simulates user behaviour.